There is never too much refactoring

+++ Labs/Lab_4_2_5(1).cpp
@@ -7,7 +7,7 @@ int main()
 	std::cout << "N: ";
 	std::cin >> n;
 
-	std::cout << n * (n + 1) / 2 << std::endl;
+	std::cout << (long double)n * (n + 1) / 2 << std::endl;
 
 	system("pause");
 	return 0;
